+++ modules/frontend/config.go
@@ -6,13 +6,13 @@ import (
 	"github.com/cortexproject/cortex/pkg/querier/frontend"
 )
 
-type FrontendConfig struct {
+type Config struct {
 	frontend.Config `yaml:",inline"`
 	QueryShards     int `yaml:"query_shards,omitempty"`
 }
 
-func (cfg *FrontendConfig) RegisterFlagsAndApplyDefaults(prefix string, f *flag.FlagSet) {
-	cfg.Config.CompressResponses = false
+func (cfg *Config) RegisterFlagsAndApplyDefaults(prefix string, f *flag.FlagSet) {
+	cfg.Config.CompressResponses = true
 	cfg.Config.DownstreamURL = ""
 	cfg.Config.LogQueriesLongerThan = 0
 	cfg.Config.MaxOutstandingPerTenant = 100